+++ lib/paper_trail/record_trail.rb
@@ -148,11 +148,19 @@ module PaperTrail
     # paper_trail-association_tracking
     def record_update_columns(changes)
       return unless enabled?
-      event = Events::Update.new(@record, false, false, changes)
+      data = Events::Update.new(@record, false, false, changes).data
+
+      # Normally, `Events::Update` copies `@record.updated_at` into
+      # `created_at`. This copy makes sense for normal updates, but
+      # update_columns does not update timestamps, so it would be incorrect
+      # here. Deleting `created_at` allows AR to set its default timestamp value
+      # instead.
+      # https://github.com/paper-trail-gem/paper_trail/pull/1396
+      data.delete(:created_at)
 
       # Merge data from `Event` with data from PT-AT. We no longer use
       # `data_for_update_columns` but PT-AT still does.
-      data = event.data.merge(data_for_update_columns)
+      data.merge!(data_for_update_columns)
 
       versions_assoc = @record.send(@record.class.versions_association_name)
       version = versions_assoc.create(data)